Split compute_instance_out.go, refactored code structure
